+++ DOMAIN/Entities/BillOfMaterials/BillOfMaterial.cs
@@ -30,7 +30,7 @@ public class BillOfMaterialItem : BaseEntity
     [StringLength(255)] public string CasNumber { get; set; }
     [StringLength(255)] public string Function { get; set; }
     public int Order { get; set; }
-    public int Quantity { get; set; } = 0; // Quantity of the component required 
+    public decimal Quantity { get; set; } = 0; // Quantity of the component required 
     public Guid? UoMId { get; set; }  // Unit of Measure, e.g., grams, liters, pieces
     public UnitOfMeasure UoM { get; set; }
     public bool IsSubstitutable { get; set; }  // Allows for substitution in production
